requirements.txt
jsonrpcclient==4.0.3
requests==2.28.2
python-dotenv==1.0.0